commit_lab10_15

+++ controllers/admin.js
@@ -1,6 +1,7 @@
 const Product = require("../models/product");
 const { validationResult } = require("express-validator/check");
 const mongoose = require("mongoose");
+const fileHelper = require("../utils/file");
 
 exports.getAddProducts = (req, res, next) => {
   res.status(422).render("admin/edit-product", {
@@ -81,6 +82,7 @@ exports.postEditProduct = (req, res, next) => {
       product.title = updatedTitle;
       product.price = updatedPrice;
       if (image) {
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = image.path;
       }
 
@@ -195,7 +197,14 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.findByIdAndRemove(prodId)
+  Product.findById(prodId)
+    .then((product) => {
+      if (!product) {
+        return next(new Error("Product not found"));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.findByIdAndRemove(prodId);
+    })
     .then((product) => {
       console.log("delete");
       res.redirect("/admin/products");
